+++ src/Screen/Layouts/Content.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Orchid\Screen;
+namespace Orchid\Screen\Layouts;
 
-use Orchid\Screen\Layouts\Base;
+use Orchid\Screen\Repository;
 
 abstract class Content extends Base
 {
